% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkConfig}
\alias{checkConfig}
\title{checkConfig}
\usage{
checkConfig(config)
}
\arguments{
\item{config}{config file with path.}
}
\value{
TRUE if config is valid.
}
\description{
checkConfig
}
\examples{
config <- system.file("extdata", "Datasets", "Demo", "info.yaml",
    package = "fourSynergy")
checkConfig(config)
}
