% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensusIa.R
\name{consensusIa}
\alias{consensusIa}
\title{consensusIa}
\usage{
consensusIa(ia, model = "F1")
}
\arguments{
\item{ia}{fourSynergy object with interactions from all base tools
(peakC, r3c-seq, fourSig, r4cker) and other relevant information.}

\item{model}{Selected optimization model. Either 'F1' or 'AUPRC'.}
}
\value{
fourSynergy object with interactions from all base tools and weighted
voting results.
}
\description{
This function performs an optimized weighted voting of 4C-seq tools.
}
\examples{
config <- system.file("extdata", "Datasets", "Demo", "info.yaml",
    package = "fourSynergy"
)
res_path <- system.file("extdata", "results", "Demo",
    package = "fourSynergy"
)
tracks <- system.file("extdata", "results", "Demo", "alignment",
    package = "fourSynergy"
)
sia <- createIa(res_path = res_path, config = config, tracks = tracks)
sia <- consensusIa(ia = sia, model = "AUPRC")
}
