% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createIa.R
\name{createIa}
\alias{createIa}
\title{createIa}
\usage{
createIa(res_path = character(), config = list(), tracks = "")
}
\arguments{
\item{res_path}{Path to results created by the pipeline.
Typically stored in the \verb{results/[dataset]/nearbait_area.bed}.}

\item{config}{Path of config file.}

\item{tracks}{Path to alignment files.}
}
\value{
fourSynergy object with interactions from all base tools.
}
\description{
This function reads the interaction bed files created by the pipeline and
transfers this information into an GrangesList.
}
\examples{
config <- system.file("extdata", "Datasets", "Demo", "info.yaml",
    package = "fourSynergy"
)
res_path <- system.file("extdata", "results", "Demo",
    package = "fourSynergy"
)
tracks <- system.file("extdata", "results", "Demo", "alignment",
    package = "fourSynergy"
)
ia <- createIa(res_path = res_path, config = config, tracks = tracks)

}
