% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karyoplots.R
\name{createKaryoplot}
\alias{createKaryoplot}
\title{Internal function to create karyoplot}
\usage{
createKaryoplot(ia, type = 1, cex = 1, cex.axis = 1, cex.lab = 1, cex.main = 1)
}
\arguments{
\item{ia}{fourSynergy object with interactions from all base tools
(peakC, r3c-seq, fourSig, r4cker) and other relevant information.}

\item{type}{Plottype.}

\item{cex}{character expansion}

\item{cex.axis}{character expansion axis}

\item{cex.lab}{character expansion labels}

\item{cex.main}{character expansion main}
}
\value{
karyoplot base
}
\description{
Internal function to create karyoplot
}
\keyword{internal}
