% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentialAnalysis.R
\name{differentialAnalysis}
\alias{differentialAnalysis}
\title{differentialAnalysis}
\usage{
differentialAnalysis(ia, fitType = "local")
}
\arguments{
\item{ia}{fourSynergy object with interactions from all base tools
(peakC, r3c-seq, fourSig, r4cker) and other relevant information.}

\item{fitType}{Parameter for DESeq2s estimateDispersions(). Should be either
"parametric", "local", "mean", or "glmGamPoi" for the type of fitting of
dispersions to the mean intensity.}
}
\value{
sia object with GRanges of DESeq results in the diff slot.
}
\description{
This function performs differential analysis to identify differential
interacting regions using DESeq2.
}
\examples{
config <- system.file("extdata", "Datasets", "Demo", "info.yaml",
    package = "fourSynergy"
)
res_path <- system.file("extdata", "results", "Demo",
    package = "fourSynergy"
)
tracks <- system.file("extdata", "results", "Demo", "alignment",
    package = "fourSynergy"
)
sia <- createIa(res_path = res_path, config = config, tracks = tracks)
sia <- consensusIa(ia = sia, model = "AUPRC")
sia <- differentialAnalysis(ia = sia, fitType = "mean")
}
\references{
https://doi.org/10.1186/s13059-014-0550-8
}
