% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fourSynergy-accessors}
\alias{fourSynergy-accessors}
\alias{getMetadata}
\alias{getExpInteractions}
\alias{getCtrlInteractions}
\alias{getExpConsensus}
\alias{getCtrlConsensus}
\alias{getViewpoint}
\alias{getVirtualFragmentLibrary}
\alias{getTracks}
\alias{getDifferential}
\alias{getDDS}
\title{fourSynergy accessors}
\usage{
getMetadata(object)

getExpInteractions(object)

getCtrlInteractions(object)

getExpConsensus(object)

getCtrlConsensus(object)

getViewpoint(object)

getVirtualFragmentLibrary(object)

getTracks(object)

getDifferential(object)

getDDS(object)
}
\arguments{
\item{object}{A \code{\linkS4class{fourSynergy}} object}
}
\value{
Various slots of the object (see individual methods)
}
\description{
Standard getter methods for \code{\linkS4class{fourSynergy}} object slots.
}
\section{Functions}{
\itemize{
\item \code{getMetadata()}: Get metadata

\item \code{getExpInteractions()}: Get experimental interactions

\item \code{getCtrlInteractions()}: Get control interactions

\item \code{getExpConsensus()}: Get experimental consensus

\item \code{getCtrlConsensus()}: Get control consensus

\item \code{getViewpoint()}: Get viewpoint

\item \code{getVirtualFragmentLibrary()}: Get virtual fragment library

\item \code{getTracks()}: Get tracks

\item \code{getDifferential()}: Get differential expression

\item \code{getDDS()}: Get DESeq results

}}
\keyword{internal}
