% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{fourSynergy-class}
\alias{fourSynergy-class}
\title{fourSynergy Class}
\description{
S4 class storing data collected from 4C-seq analyses.
}
\section{Slots}{

\describe{
\item{\code{metadata}}{Experimental metadata from config file.}

\item{\code{expInteractions}}{Base tool interactions found in the experiment.}

\item{\code{ctrlInteractions}}{Base tool interactions found in the control.}

\item{\code{expConsensus}}{Consensus interactions found in the experiment.}

\item{\code{ctrlConsensus}}{Consensus interactions found in the control.}

\item{\code{vp}}{Viewpoint position.}

\item{\code{vfl}}{Virtual fragment library.}

\item{\code{tracks}}{Path to the alignment files.}

\item{\code{differential}}{Results of differential interaction calling (res).}

\item{\code{dds}}{Results of differential interaction calling (dds).}
}}

