% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fourSynergy.R
\docType{package}
\name{fourSynergy-package}
\alias{fourSynergy}
\alias{fourSynergy-package}
\title{fourSynergy: Ensemble based interaction calling in 4C-seq data}
\description{
fourSynergy is an ensemble algorithm leveraging synergies among the existing 4C-seq algorithms r3C-seq, peakC, r.4cker and fourSig. It uses a weighted voting approach to perform improved interaction calling. fourSynergy supports also differential interaction calling.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sophiewind/fourSynergy}
  \item Report bugs at \url{https://github.com/sophiewind/fourSynergy/issues}
}

}
\author{
\strong{Maintainer}: Sophie Wind \email{sophie.wind@uni-muenster.de} (\href{https://orcid.org/0009-0002-1174-8201}{ORCID})

Authors:
\itemize{
  \item Carolin Walter \email{carolin.walter@uni-muenster.de} (\href{https://orcid.org/0000-0002-5155-7728}{ORCID}) [funder]
}

}
