% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fourSynergy-setters}
\alias{fourSynergy-setters}
\alias{setExpConsensus}
\alias{setCtrlConsensus}
\alias{setDifferential}
\alias{setDDS}
\alias{setDds}
\alias{setctrlConsensus}
\title{fourSynergy setters}
\usage{
setDifferential(object, value)

setDds(object, value)

setExpConsensus(object, value)

setctrlConsensus(object, value)
}
\arguments{
\item{object}{A \code{\linkS4class{fourSynergy}} object}

\item{value}{Replacement value for the respective slot}
}
\value{
Updated \code{fourSynergy} object (invisibly)
}
\description{
Standard setter methods for \code{\linkS4class{fourSynergy}} object slots.
}
\section{Functions}{
\itemize{
\item \code{setDifferential()}: Set setDifferential slot

\item \code{setDds()}: Set setDds slot

\item \code{setExpConsensus()}: Set setExpConsensus slot

\item \code{setctrlConsensus()}: Set setCtrlConsensus slot

}}
\keyword{internal}
