% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBaseTracks.R
\name{plotBaseTracks}
\alias{plotBaseTracks}
\title{plotBaseTracks}
\usage{
plotBaseTracks(sia, highlight_regions = NULL, max_range = 3000)
}
\arguments{
\item{sia}{fourSynergy object with interactions from all base tools}

\item{highlight_regions}{regions to highlight in the plot}

\item{max_range}{maximum plotting range}
}
\value{
Track-plots for all treatments with interactions from base tools
}
\description{
plotBaseTracks
}
\examples{
config <- system.file("extdata", "Datasets", "Demo", "info.yaml",
    package = "fourSynergy"
)
res_path <- system.file("extdata", "results", "Demo",
    package = "fourSynergy"
)
tracks <- system.file("extdata", "results", "Demo", "alignment",
    package = "fourSynergy"
)
sia <- createIa(res_path = res_path, config = config, tracks = tracks)
plotBaseTracks(sia)
}
