% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotConsensusIa.R
\name{plotConsensusIa}
\alias{plotConsensusIa}
\title{plotConsensusIa}
\usage{
plotConsensusIa(
  ia = GRangesList(),
  genes_of_interest = NULL,
  cex.chr = 1,
  cex.ideo = 0.6,
  cex.y.lab = 0.6,
  cex.y.track = 0.6,
  cex.vp = 1,
  cex.leg = 0.6,
  highlight_regions = NULL,
  plot_spider = FALSE,
  gene.name.cex = 1
)
}
\arguments{
\item{ia}{fourSynergy object with interactions from all base tools
(peakC, r3c-seq, fourSig, r4cker) and other relevant information.}

\item{genes_of_interest}{Vector with genes of interest.}

\item{cex.chr}{character expansion of chromosome label.}

\item{cex.ideo}{character expansion base numbers of ideogram.}

\item{cex.y.lab}{character expansion for y labels.}

\item{cex.y.track}{character expansion y axis track.}

\item{cex.vp}{character expansion viewpoint label.}

\item{cex.leg}{character expansion for legend.}

\item{highlight_regions}{regions to highlight in the plot.}

\item{plot_spider}{plotting connections from VP to interactions.}

\item{gene.name.cex}{character expansion for gene names.}
}
\value{
karyoplot with calling results.
}
\description{
This function creates a karyotype plot displaying the interaction calls from
the consensus approach.
}
\examples{
config <- system.file("extdata", "Datasets", "Demo", "info.yaml",
    package = "fourSynergy"
)
res_path <- system.file("extdata", "results", "Demo",
    package = "fourSynergy"
)
tracks <- system.file("extdata", "results", "Demo", "alignment",
    package = "fourSynergy"
)
sia <- createIa(res_path = res_path, config = config, tracks = tracks)
sia <- consensusIa(ia = sia, model = "AUPRC")
plotConsensusIa(ia = sia)
}
