% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIaIndividualTools.R
\name{plotIaIndiviualTools}
\alias{plotIaIndiviualTools}
\title{plotIaIndiviualTools}
\usage{
plotIaIndiviualTools(
  ia = GRangesList(),
  genes_of_interest = NULL,
  cex.chr = 1,
  cex.ideo = 0.6,
  cex.y.track = 0.6,
  cex.y.lab = 0.6,
  cex.vp = 1
)
}
\arguments{
\item{ia}{fourSynergy object with interactions from all base tools
(peakC, r3c-seq, fourSig, r4cker) and other relevant information.}

\item{genes_of_interest}{Vector with genes of interest.}

\item{cex.chr}{character expansion of chromosome label.}

\item{cex.ideo}{character expansion base numbers of ideogram.}

\item{cex.y.track}{character expansion y axis track.}

\item{cex.y.lab}{character expansion y lab.}

\item{cex.vp}{character expansion viewpoint label.}
}
\value{
karyoplot with calling results.
}
\description{
This function creates a karyoplot with the interactions calls of the
individual tools.
}
\examples{

config <- system.file("extdata", "Datasets", "Demo", "info.yaml",
    package = "fourSynergy"
)
res_path <- system.file("extdata", "results", "Demo",
    package = "fourSynergy"
)
tracks <- system.file("extdata", "results", "Demo", "alignment",
    package = "fourSynergy"
)
sia <- createIa(res_path = res_path, config = config, tracks = tracks)
plotIaIndiviualTools(ia = sia, genes_of_interest = c("Ldlrad4", "Cep76"))
}
