% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plotpreTracks}
\alias{plotpreTracks}
\title{plotpreTracks}
\usage{
plotpreTracks(sia, highlight_regions = NULL)
}
\arguments{
\item{sia}{fourSynergy object with interactions from all base tools.}

\item{highlight_regions}{regions to highlight in the plot.}
}
\value{
list with read counts and Granges of bedfiles
}
\description{
plotpreTracks
}
\keyword{internal}
