% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karyoplots.R
\name{plotRegions}
\alias{plotRegions}
\title{Internal function to highlight regions in karyoplot.}
\usage{
plotRegions(ia, kp, highlight_regions)
}
\arguments{
\item{ia}{fourSynergy object with interactions from all base tools
(peakC, r3c-seq, fourSig, r4cker) and other relevant information.}

\item{kp}{karyoplot.}

\item{highlight_regions}{regions to highlight in the plot.}
}
\value{
karyoplot base
}
\description{
Internal function to highlight regions in karyoplot.
}
\keyword{internal}
