% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karyoplots.R
\name{plotTracks}
\alias{plotTracks}
\title{Internal function to plot tracks}
\usage{
plotTracks(ia, kp, bgs, r0 = 0, r1 = 1, cex.vp = 1, cex.y.track = 0.6)
}
\arguments{
\item{ia}{fourSynergy object with interactions from all base tools
(peakC, r3c-seq, fourSig, r4cker) and other relevant information.}

\item{kp}{Kayroplot object.}

\item{bgs}{GrangesList of bedGraphs.}

\item{r0}{r0 defines the vertical range of the data panel.}

\item{r1}{r1 defines the vertical range of the data panel.}
}
\value{
karyoplot with tracks
}
\description{
Internal function to plot tracks
}
\keyword{internal}
