% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readAndTag}
\alias{readAndTag}
\title{readAndTag}
\usage{
readAndTag(file_path, tag, org)
}
\arguments{
\item{file_path}{Path to the .bed files
(\verb{results/[dataset]/nearbait_area.bed}).}

\item{tag}{tool name.}

\item{org}{organism.}
}
\value{
GRanges with interaction calls and tool name as mcol.
}
\description{
Internal function to read base tools solutions in FourSynergy format.
}
\keyword{internal}
