% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/karyoplots.R
\name{readBedGraph}
\alias{readBedGraph}
\title{Internal function to read bedGraphs}
\usage{
readBedGraph(ia)
}
\arguments{
\item{ia}{fourSynergy object with interactions from all base tools.
(peakC, r3c-seq, fourSig, r4cker) and other relevant information.}
}
\value{
GrangesList of bedGraph content
}
\description{
Internal function to read bedGraphs
}
\keyword{internal}
