% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{setDds}
\alias{setDds}
\title{Set dds attribute}
\usage{
setDds(object, value)
}
\arguments{
\item{object}{fourSynergy object with interactions from all base tools.}

\item{value}{DESeqDataSet-object.}
}
\value{
Modified fourSynergy-object
}
\description{
Set dds attribute
}
