% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{setDifferential,fourSynergy-method}
\alias{setDifferential,fourSynergy-method}
\title{Set differential attribute}
\usage{
\S4method{setDifferential}{fourSynergy}(object, value)
}
\arguments{
\item{object}{fourSynergy object with interactions from all base tools}

\item{value}{DESeqResults-object}
}
\value{
Modified fourSynergy-object
}
\description{
Set differential attribute
}
