% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDifferentialIa.R
\name{plotDiffIa}
\alias{plotDiffIa}
\title{plotDiffIa}
\usage{
plotDiffIa(
  ia,
  genes_of_interest = NULL,
  cex.chr = 1,
  cex.ideo = 0.6,
  cex.y.track = 0.6,
  cex.vp = 1
)
}
\arguments{
\item{ia}{fourSynergy object with interactions from all base tools
(peakC, r3c-seq, fourSig, r4cker) and more relevant information.}

\item{genes_of_interest}{Vector with genes of interest.}

\item{cex.chr}{character expansion of chromosome label.}

\item{cex.ideo}{character expansion base numbers of ideogram.}

\item{cex.y.track}{character expansion y axis track.}

\item{cex.vp}{character expansion viewpoint label.}
}
\value{
DESeq2 results of differential interaction calling.
}
\description{
This function creates a karyoplot with the differential interactions calls.
}
\examples{
config <- system.file("extdata", "Datasets", "Demo", "info.yaml",
    package = "fourSynergy"
)
res_path <- system.file("extdata", "results", "Demo",
    package = "fourSynergy"
)
tracks <- system.file("extdata", "results", "Demo", "alignment",
    package = "fourSynergy"
)
sia <- createIa(res_path = res_path, config = config, tracks = tracks)
sia <- consensusIa(ia = sia, model = "AUPRC")
sia <- differentialAnalysis(ia = sia, fitType = "mean")
plotDiffIa(ia = sia)
}
